backend/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-stubs==4.2.3
django-stubs-ext==4.2.2
graphql-core==3.2.3
mypy==1.4.1
mypy-extensions==1.0.0
Pillow==10.0.0
PyJWT==2.7.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
strawberry-django-auth==0.375.0
strawberry-graphql==0.194.4
strawberry-graphql-django==0.11.0
tomli==2.0.1
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
typing_extensions==4.7.1
